Refactor mergePackageJson, move to another file
