+++ packages/cli/src/lib/file.ts
@@ -1,11 +1,16 @@
-import {resolvePath} from '@shopify/cli-kit/node/path';
+import {readdir} from 'node:fs/promises';
+import {resolvePath, joinPath} from '@shopify/cli-kit/node/path';
 import {
   readFile,
   writeFile,
   fileExists,
   isDirectory,
 } from '@shopify/cli-kit/node/fs';
-import {readdir} from 'fs/promises';
+import {
+  readAndParsePackageJson,
+  writePackageJSON,
+  type PackageJson as _PackageJson,
+} from '@shopify/cli-kit/node/node-package-manager';
 import {formatCode, type FormatOptions} from './format-code.js';
 
 export async function replaceFileContent(
@@ -59,4 +64,83 @@ export async function findFileWithExtension(
   }
 
   return {};
+}
+
+type PackageJson = _PackageJson & {
+  peerDependencies?: _PackageJson['dependencies'];
+  comment?: string;
+};
+
+const MANAGED_PACKAGE_JSON_KEYS = Object.freeze([
+  'dependencies',
+  'devDependencies',
+  'peerDependencies',
+] as const);
+
+type ManagedKey = (typeof MANAGED_PACKAGE_JSON_KEYS)[number];
+
+export async function mergePackageJson(
+  sourceDir: string,
+  targetDir: string,
+  options?: {ignoredKeys?: string[]},
+) {
+  const targetPkgJson: PackageJson = await readAndParsePackageJson(
+    joinPath(targetDir, 'package.json'),
+  );
+  const sourcePkgJson: PackageJson = await readAndParsePackageJson(
+    joinPath(sourceDir, 'package.json'),
+  );
+
+  const ignoredKeys = new Set(['comment', ...(options?.ignoredKeys ?? [])]);
+
+  const unmanagedKeys = Object.keys(sourcePkgJson).filter(
+    (key) => !MANAGED_PACKAGE_JSON_KEYS.includes(key as ManagedKey),
+  ) as Exclude<keyof PackageJson, ManagedKey>[];
+
+  for (const key of unmanagedKeys) {
+    if (ignoredKeys.has(key)) continue;
+
+    const sourceValue = sourcePkgJson[key];
+    const targetValue = targetPkgJson[key];
+
+    const newValue =
+      Array.isArray(sourceValue) && Array.isArray(targetValue)
+        ? [...targetValue, ...sourceValue]
+        : typeof sourceValue === 'object' && typeof targetValue === 'object'
+        ? {...targetValue, ...sourceValue}
+        : sourceValue;
+
+    targetPkgJson[key] = newValue as any;
+  }
+
+  const remixVersion = Object.entries(targetPkgJson.dependencies || {}).find(
+    ([dep]) => dep.startsWith('@remix-run/'),
+  )?.[1];
+
+  for (const key of MANAGED_PACKAGE_JSON_KEYS) {
+    if (ignoredKeys.has(key)) continue;
+
+    if (sourcePkgJson[key]) {
+      targetPkgJson[key] = [
+        ...new Set([
+          ...Object.keys(targetPkgJson[key] ?? {}),
+          ...Object.keys(sourcePkgJson[key] ?? {}),
+        ]),
+      ]
+        .sort()
+        .reduce((acc, dep) => {
+          let version = (sourcePkgJson[key]?.[dep] ??
+            targetPkgJson[key]?.[dep])!;
+
+          if (dep.startsWith('@remix-run/') && remixVersion) {
+            version = remixVersion;
+          }
+
+          acc[dep] = version;
+          return acc;
+        }, {} as Record<string, string>);
+    }
+  }
+
+  await writePackageJSON(targetDir, targetPkgJson);
 }
